refactor: 話者単位の合成初期化を tts_pipeline router へ移動
